// apps/admin/src/lib/api/generated/models/accountInputTier.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.51
 */

export type AccountInputTier = typeof AccountInputTier[keyof typeof AccountInputTier];


export const AccountInputTier = {
  bronze: 'bronze',
  silver: 'silver',
  gold: 'gold',
  platinum: 'platinum',
  diamond: 'diamond',
} as const;